Properties should be imutable

+++ src/lib/gui/button.cpp
@@ -11,7 +11,7 @@ Button::Button()
 Button *Button::layout()
 {
 	std::lock_guard<std::recursive_mutex> lock(m_mutex);
-	auto border = (int)get_long_prop("border");
+	auto border = get_long_prop("border");
 	auto pos = border;
 	if (m_state != 1) pos *= 2;
 	m_flow->change(pos, pos, m_w - border * 2, m_h - border * 2)->layout();
@@ -23,7 +23,7 @@ Button *Button::draw(const Ctx &ctx)
 {
 	//allready locked by GUI thread
 	auto col = (uint32_t)get_long_prop("color");
-	auto border = (int)get_long_prop("border");
+	auto border = get_long_prop("border");
 	ctx.panel(col, true, border * m_state, 0, 0, m_w, m_h);
 	return this;
 }
