+++ src/lib/gui/slider.h
@@ -13,7 +13,7 @@ public:
 	Slider *mouse_up(const std::shared_ptr<Msg> &event) override;
 	Slider *mouse_move(const std::shared_ptr<Msg> &event) override;
 	int m_state = 0;
-	int m_old_value = 0;
+	int64_t m_old_value = 0;
 	int m_down_xy = 0;
 };
